Convert AI model docs UI from React to native Astro (#32343)
* Convert AI model docs UI from React to native Astro

Breaks out the UI-only changes from PR #30505 so the middlecache
PR can focus solely on data loading.

Catalog filter controller:
- Replace the large ModelFilters.tsx React island (449 lines) with
  a vanilla JS <script> that owns filter/sort/URL state and re-flows
  the corner-mark grid — no React state management needed
- Split the old island into small, self-contained React islands:
  FilterDropdown/FilterDropdownWrapper (multi-select) and
  SortSelect/SortSelectWrapper (sort order) that fire CustomEvents
- All islands restyled to Nimbus design tokens (border-border, bg-card,
  text-foreground, etc.) matching the old island's styling

Model detail page:
- Replace plain <h2 id> section headings with AnchorHeading so they
  get the .heading-wrapper anchor link structure from the rehype pipeline
- Extract inline Open/Download SVGs into SchemaFileLinks.astro

SchemaDisplay:
- Replace Date.now() schemaId with a stable hash of the schema JSON
  so sessionStorage expand/collapse state persists across builds

ModelFeatures:
- Fix stale links: /workers-ai/glossary/ → /workers-ai/platform/glossary/
- Fix stale link: /workers-ai/function-calling → /workers-ai/function-calling/

* Address code review findings (CR-446f, CR-44f4, CR-2232, CR-d5d3, CR-781f, CR-fa02, CR-1cec, CR-ce42)

Security:
- Replace set:html JSON injection in <script> with data-* attribute
  (Astro escapes attribute values, preventing </script> breakout)

Accessibility:
- Add aria-label to SchemaFileLinks open/download icon-only links
- Add aria-label to function-calling external link in ModelFeatures

Correctness:
- Lowercase heading slugs in headings array + AnchorHeading calls so
  TOC anchor links match the IDs github-slugger produces
- Remove false CustomEvent claim from SortSelect.tsx JSDoc (the
  wrapper dispatches events, not the presentational component)
- Remove leading space in countLabelEl.textContent (doubled with
  the existing {" "} separator)
- Replace URLSearchParams.size with toString() === '' for older
  browser compatibility

Code quality:
- Extract shared link class string to const in SchemaFileLinks.astro
