+++ src/components/models/ModelCard.astro
@@ -6,7 +6,7 @@
  * a 12px meta row (author • model type), 14px muted description, badges.
  *
  * The cell carries the `data-facet-*` / `data-name` / `data-date` attributes the
- * catalog controller (ModelFilters.tsx island) reads to filter, sort, and re-flow
+ * catalog controller (vanilla JS script) reads to filter, sort, and re-flow
  * the corner marks. `index`/`cols` seed the build-time corner marks.
  */
 import { Icon } from "astro-icon/components";
