+++ src/components/models/ModelFeatures.astro
@@ -82,7 +82,7 @@ const price = Array.isArray(properties.price)
 					<tr>
 						<td>
 							Context Window
-							<a href="/workers-ai/glossary/">
+							<a href="/workers-ai/platform/glossary/">
 								<span class="external-link"> ↗</span>
 							</a>
 						</td>
@@ -125,7 +125,10 @@ const price = Array.isArray(properties.price)
 					<tr>
 						<td>
 							Function calling{" "}
-							<a href="/workers-ai/function-calling">
+							<a
+								href="/workers-ai/function-calling/"
+								aria-label="Function calling"
+							>
 								<span class="external-link"> ↗</span>
 							</a>
 						</td>
