+++ src/components/models/SchemaFileLinks.astro
@@ -1,0 +1,68 @@
+---
+/**
+ * SchemaFileLinks — renders Open and Download icon buttons for a schema file.
+ * Extracted from ModelDetailPage to avoid repeating inline SVGs.
+ */
+
+interface Props {
+	href: string;
+	filename: string;
+}
+
+const { href, filename } = Astro.props;
+
+const linkClass =
+	"text-muted-foreground hover:bg-accent hover:text-foreground flex h-7 w-7 items-center justify-center rounded no-underline";
+---
+
+<span class="flex items-center gap-1">
+	<a
+		href={href}
+		target="_blank"
+		rel="noopener noreferrer"
+		title="Open"
+		aria-label={`Open ${filename}`}
+		class={linkClass}
+	>
+		<svg
+			xmlns="http://www.w3.org/2000/svg"
+			width="14"
+			height="14"
+			viewBox="0 0 24 24"
+			fill="none"
+			stroke="currentColor"
+			stroke-width="2"
+			stroke-linecap="round"
+			stroke-linejoin="round"
+			aria-hidden="true"
+		>
+			<path d="M18 13v6a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V8a2 2 0 0 1 2-2h6"></path>
+			<polyline points="15 3 21 3 21 9"></polyline>
+			<line x1="10" y1="14" x2="21" y2="3"></line>
+		</svg>
+	</a>
+	<a
+		href={href}
+		download={filename}
+		title="Download"
+		aria-label={`Download ${filename}`}
+		class={linkClass}
+	>
+		<svg
+			xmlns="http://www.w3.org/2000/svg"
+			width="14"
+			height="14"
+			viewBox="0 0 24 24"
+			fill="none"
+			stroke="currentColor"
+			stroke-width="2"
+			stroke-linecap="round"
+			stroke-linejoin="round"
+			aria-hidden="true"
+		>
+			<path d="M21 15v4a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2v-4"></path>
+			<polyline points="7 10 12 15 17 10"></polyline>
+			<line x1="12" y1="15" x2="12" y2="3"></line>
+		</svg>
+	</a>
+</span>
